sites_visited: Reorder computations for efficiency
Dask doesn't have a good query planner (the philosophy is that it
tries to not make assumptions???), so you as the programmer need to
mentally keep track of the cardinality of operations to keep the frame
sizes from exploding. Since grouby is a reduction, push it early into
the query graph as possible.

Importantly, "melt" is a passthrough function to pandas and will
create a single output partition. If melt is called on the dask
dataframe it will get pushed early into the compute graph since it
doesn't have any requirements! You can see this by watching the
computation run on 8787 and learning to read dask tarot cards. Early
in the graph though it gets too many inputs into one output, and it
will melt itself (aka take the entire dataset in RAM)! This commit
moves the melt operation after computing the groupby so it will only
operate on the small resultant frame.