Remove unnecessary rename call
This was not doing what the original author thought it was doing. The
rename was applying to the _rows_ and setting the first entry to
"user". This only showed up as an error now since there is no longer
an off by one error in cohort assignment and cohort 0 is valid.