Convert category to a faceted chart
One tricky thing is that the data needs to be specified at the facet
level, not with the area chart objects directly. Specifying directly
results in both charts just using the first column (downlink). This
might be avoidable with the altair chart shorthand using column= in
the area chart, but for now this is working : )